Clean up dead files and duplicate UI plumbing
