--- src/components/MobileSidebar.tsx
@@ -1,9 +1,0 @@
-"use client";
-
-import SidebarContent from "./SidebarContent";
-
-export default function MobileSidebar({ onNavigate }: { onNavigate: () => void }) {
-  return <SidebarContent onNavigate={onNavigate} />;
-}
-
-
